refactor(ask-user-question): dedupe inline-input draft helpers; document chat invariants
Address PR #1337 review feedback (claude[bot]):

- Hoist the duplicated withInlineDraft and the near-duplicate per-owner draft/caret accessors (inlineDraftAt/inlineCaretAt vs inlineDraftValue/inlineCaretValue) out of state-reducer.ts and questionnaire-session.ts into a single shared state/inline-input.ts module (readInlineDraft, readInlineCaret, withInlineDraft, resolveInlineDraftValue), keeping the owner-branching logic in one place.

- Document why the chatFocused branch needs the Left/Right ignore guard while the 'other' inline editor does not (chat falls through to tabSwitchAction; other does not).

- Document the chatAnswerValue live-buffer invariant (reads runtime.inputBuffer because confirm only fires while chat owns the live editor; returns untrimmed text).

- Document the mixed-dialog precedence: signal-only chat wins the terminate path even when another question carried typed chat (deliberate product decision).

Refs: #1331
Assistant-model: Claude Opus 4.8
