+++ packages/coding-agent/src/core/tools/ask-user-question/state/key-router.ts
@@ -58,6 +58,14 @@ function computeAutoAdvanceTab(state: QuestionnaireState, runtime: Questionnaire
 	return runtime.questions.length;
 }
 
+/**
+ * Resolve the chat answer text at confirm time. Reads the LIVE `runtime.inputBuffer`
+ * rather than a persisted draft because confirm only reaches here while chat is the
+ * focused row and therefore owns the live inline editor (`inlineInputOwner === "chat"`),
+ * so the in-flight buffer is authoritative. The blank-vs-typed decision uses `trim()`,
+ * but the returned text is left UNTRIMMED to preserve the user's literal message; a
+ * blank/whitespace buffer falls back to the reserved sentinel label (signal-only chat).
+ */
 function chatAnswerValue(state: QuestionnaireState, runtime: QuestionnaireRuntime, sentinelLabel: string): string {
 	if (state.inlineInputOwner !== "chat") return sentinelLabel;
 	return runtime.inputBuffer.trim().length > 0 ? runtime.inputBuffer : sentinelLabel;
@@ -177,6 +185,13 @@ export function routeKey(data: string, state: QuestionnaireState, runtime: Quest
 		if (kb.matches(data, KEYBIND_DOWN)) {
 			return { kind: "focus_options", optionIndex: 0 };
 		}
+		// Left/Right guard is intentionally asymmetric — it exists for the chat owner but
+		// NOT the "other" inline editor. The chatFocused branch falls through to
+		// `tabSwitchAction` below, which in multi-question mode consumes Left/Right for tab
+		// switching; swallowing them here keeps Left/Right as in-editor caret movement for
+		// the chat input. The "other" inline editor flows through the separate
+		// `state.inputMode` branch (further down), which never calls `tabSwitchAction`, so it
+		// needs no equivalent guard. Do not "simplify" by unifying the two paths.
 		if (
 			state.inputMode &&
 			state.inlineInputOwner === "chat" &&
